Handle issues with joining channels.
This will start a query with whoever invited the bot based on whatever
issue the bot ran into. In the case of a key being required to join the
channel, the bot will start a dialog with the inviter to ask for a key.
In the case of being banned or the channel being full, the attempt to
join will be immediately abandoned.

This fixes #16 - Bot can not join channels with keys.